Big restructure of map layout
