+++ src/main.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 import { BrowserRouter } from 'react-router-dom'
-import './index.css'
+import './css/index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <BrowserRouter>
